Apparently GraphQL didn't like the specificity
