+++ test/Datadog.Trace.ClrProfiler.IntegrationTests/Helpers/AspNetCoreMvcSpanExpectation.cs
@@ -1,3 +1,5 @@
+using Datadog.Trace.TestHelpers;
+
 namespace Datadog.Trace.ClrProfiler.IntegrationTests
 {
     public class AspNetCoreMvcSpanExpectation : WebServerSpanExpectation
@@ -5,6 +7,17 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests
         public AspNetCoreMvcSpanExpectation(string serviceName, string operationName)
             : base(serviceName, operationName)
         {
+        }
+
+        public override bool Matches(MockTracerAgent.Span span)
+        {
+            var spanUri = GetTag(span, Tags.HttpUrl);
+            if (spanUri == null || !spanUri.Contains(OriginalUri))
+            {
+                return false;
+            }
+
+            return base.Matches(span);
         }
     }
 }
